Работа над чтением файлов
1) Добавил использование абстрактного класса для чтения файлов
2) Добавил названия к измерениям на графики (в хранилище данных графика и реализациях графика)

+++ csvreader.cpp
@@ -18,7 +18,7 @@ bool CsvReader::readData(const QString filePath, ChartData &data, QString &readE
     ChartDataPoint dataPoint;
     foreach (const QString & value, csvData) {
         if (index) {
-            dataPoint.value = value;
+            dataPoint.value = value.toFloat();
             data.points.push_back(dataPoint);
             index = !index;
         } else {
